[TEMP] Restore old behavior where "Sans" was used as fallback font
This patch should be removed in next rebase cycle.

https://codereview.chromium.org/31923005 changed the fallback font logic,
it was probably not intended since it merely moved code around.
This restores the logic so that "Sans" is always tried when looking for
a fallback family.

SkFontConfigInterface_direct.cpp looks for fallbacks among all
available fonts iff "Sans", "Serif", "Monospace" or "" is requested.
So it's important that at least one of those is tried.

This avoids crashing the renderer process if "Arial" is missing on
the system.

BUG=309714
AUTHOR=rogerj@opera.com
TBR=eae@chromium.org

Review URL: https://codereview.chromium.org/64243002

git-svn-id: svn://svn.chromium.org/blink/trunk@161534 bbb929c8-8fbe-4397-9dbb-9b2b20218538