Do not set zoom: auto to the actual minimum possible scale
But to the one calculated from the intented layout size. The reasoning
is that it might be possible to zoom the page out further that they
layout width in the case components on the page grow beyond this width.

That is an issue as it can cause different scaled when navigating subpages.

https://code.google.com/p/chromium/issues/detail?id=326149