Fix for crash in Canvas2DLayerBridge related to graphics driver reset
After a GPU reset it may not be possible to obtain a graphics context 
provider object. What this means is that it is only safe to assume that 
m_contextProvider is non-null when m_surfaceIsValid is true. 

This change swaps the operands of logical operators in order to benefit 
from predicate short-circuiting on m_surfaceIsValid, and it adds a 
simple null pointer check.

BUG=366719

Review URL: https://codereview.chromium.org/251493007/



git-svn-id: svn://svn.chromium.org/blink/branches/chromium/1916@172634 bbb929c8-8fbe-4397-9dbb-9b2b20218538