Regression: Transition of image moves from its actual position on clicking 'Multi-Flip' option.
This CL removes the partial graphics layer geometry update code path. Now we
always just a full-tree geometry update. That's a bit sad, but it's what we
used to do in M34. On trunk, we still do partial updates (hopefully correctly).

Unfortunately, we aren't able to test this CL on trunk because this code path
works differently now, which is the same reason we can't merge back the fix for
the multi-flip example.

BUG=359940


git-svn-id: svn://svn.chromium.org/blink/branches/chromium/1916@172807 bbb929c8-8fbe-4397-9dbb-9b2b20218538