Chromium CPU profiler: source line support
This is the second and final part of adding source line support in Chromium CPU profiler that needs to be done to move XDK CPU profiler on the same way as CDT uses.

In brief, we need to do two things:
1) Map the samples to source lines during profiling session and add new V8 public API to get the hit source lines. These changes are already merged into v8-crosswalk.
2) The Chrome profilers works as follows: Chrome Browser ==(a "start/stop profiling" command)==> WebKit (WebSocket) ==> WebKit(InspectorBackendDispatcher) =="start/stop profiling"==> WebKit(InspectorProfilerAgent) ==V8 Public APIs (StartProfiling/Stopprofiling)==> V8.
So, on StopProfiling command, WebKit(InspectorProfilerAgent) gets the collected results (using V8 public API) and packs them into JSON to send back to Chrome browser.

This patch implements writing info about the hit source lines in JSON data to be sent back to Chromium browser.