Add imageFilenameExtension API to WebElement
ChromiumEFL relies on a WebCore::Image API being exposed
in order to populate some context menu items.

Namely WebCore::Image::fileNameExtension();

Although there is a WebImage public class/API
that in theory could be extended, WebImage is not
about images stored in a file format, but mainly
about the pixel data of an image (for instance,
WebImage wraps a SkBitmap).

Because of this, it is preferable to extend WebElement accordingly.

A special note about the file name: Technically, "Filename"
should be camel cased, but I left it as-is to match the
existing Image counterpart API.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2547