Revert "[Backport Patch] Use glDiscardFramebuffer instead of glClear"
There was a follow-up to this patch upstream in crrev.com/656053002
which reverted the performance gains originally obtained.

XWALK-2719 and the original pull request #25 contains more information
about what happened, and crrev.com/667733002 is fixing it in a different
way.

For now, let's wait for everything to settle upstream and go back to the
original M39 behavior.

This reverts commit 90a30f1f6273a6b6bf25263c18ba453c99b6f570.

BUG=XWALK-2719