[Android] Enable WebCL on devices which only have OpenCL 1.1 support.
WebCL 1.0 is based on OpenCL 1.1, so this feature should be available as
long as the device supports OpenCL 1.1 or above. This commit eliminates
the hard dependencies on several OpenCL 1.2 APIs.

WebCLDevice::release() is deliberately removed since clReleaseDevice() is only
available since OpenCL 1.2 to release a sub device, while it is actually not
required by the WebCL 1.0 implementation.

BUG=XWALK-3975