bindings: Do not use ScriptWrappable::wrap for RemoteDOMWindow (2nd time).
Re-apply the same fix as http://crrev.com/1100693002 .
See the description of that CL for details.

The difference between this second CL and the first CL is that
this CL keeps the same interface name "LocalDOMWindow" and does not rename
it to "DOMWindow" because
1) Chromium's browser_tests ActivityLogApiTest.TriggerEvent depends on
   the interface name (hard-coded), and
2) We'd like to apply this patch to M43 as "merge", so we'd like to avoid
   3-way commits.

Thus, this CL keeps the same name "LocalDOMWindow".  However, I don't like
using this name for RemoteDOMWindow (it's confusing), so I'll rename it in
separate CLs (3-way commits), which do not need to be merged to M43.

BUG=478890,477410,475880

Review URL: https://codereview.chromium.org/1098233003

(cherry picked from commit d1f1e636870ea4188b626abc0576571dd6303608)

Review URL: https://codereview.chromium.org/1088933008

git-svn-id: svn://svn.chromium.org/blink/branches/chromium/2357@194359 bbb929c8-8fbe-4397-9dbb-9b2b20218538