Leak Detector: Always create ScriptRegexp to count objects accurately
V8PerContextData for ScriptRegexp is kept even after refreshing a
page, and this is initialized e.g. when isValidEmailAddress is called.
this means the number of V8PerContextData after loading a test
depends if the test calls isValidEmailAddress or something. This CL
has the leak detector call ensureScriptRegexpContext and ensures
that V8PerContextData is always created.

By this CL, the result of leak detection of V8PerContextData will be
more reliable.

I plan to fix the Chromium side and LeakExpectations later.

Chromium side: crrev.com/1235813002

BUG=506466, 508453

Review URL: https://codereview.chromium.org/1231413002

git-svn-id: svn://svn.chromium.org/blink/trunk@198759 bbb929c8-8fbe-4397-9dbb-9b2b20218538