Perform layout of subtree roots by depth
When subtree roots are nested, it's possible for a positioned
object to be inside one, but laid out by its containing block
in another. In this case, if it's new, we need to lay out the
descendant subtree first to add it to its ancestor's tracked
descendant map before laying out the ancestor subtree.

This patch moves the HashSet of subtrees into a class that
returns them in an ordered vector for iteration instead of
the arbitrary ordering in HashSet.

BUG=499199

Review URL: https://codereview.chromium.org/1229893005

git-svn-id: svn://svn.chromium.org/blink/trunk@198898 bbb929c8-8fbe-4397-9dbb-9b2b20218538