May need to insert or remove column sets when out-of-flow objects get their containing block changed.
Out-of-flow descendants may become or cease to be a part of a multicol
container due to changes on an ancestor. An ancestor may become or cease to be
a containing block for out-of-flow descendants.

Transforms also establish containing blocks for absolutely and fixed positioned
descendants, so we need to check for that too.

BUG=509463
R=dsinclair@chromium.org

Review URL: https://codereview.chromium.org/1229983004

git-svn-id: svn://svn.chromium.org/blink/trunk@198940 bbb929c8-8fbe-4397-9dbb-9b2b20218538