Don't underflow FiringEventIterator::iterator when removing listeners.
When an EventTarget has multiple EventListeners for an event type, the
target iterates over the listener and fires them in order. If a script
removes a listener during the iteration, the iterator position may
have to be updated.

Previously, EventTarget::fireEventListeners positioned its iterator at
the firing event listener when firing the event listener. However
because the iterator is a size_t index (with minimum value zero), it
cannot point "before" the first listener. Thus a situation can arise
where, if the first listener has not fired but is removed, the
iterator underflows. This had the incorrect but thankfully innocuous
behavior of terminating the iteration too early.

This change makes fireEventListeners position its iterator at the
listener after the currently firing iterator (if any, or the end
otherwise) so that it is always possible to move the iterator back one
if necessary. This is sufficient, even if the first iterator is
successively removed, because subsequent "first" iterators are then at
the "next" position (0) which does not require updating the iterator.

BUG=262570

Review URL: https://codereview.chromium.org/1235803009

git-svn-id: svn://svn.chromium.org/blink/trunk@198946 bbb929c8-8fbe-4397-9dbb-9b2b20218538