Fix for building partially-specified crop rects.
When building crop rects for Skia, and only the left (or top) edge
of the effect boundaries is set, we no longer need to force the
setting of the right (or bottom) edge: Skia will compute the opposite
edge from the rect formed from the union of the filter's inputs. We
should not be falling back to the filter region in this case.

Covered by existing tests (e.g., svg/batik/filters/filterRegions.svg).

Review URL: https://codereview.chromium.org/1238753004

git-svn-id: svn://svn.chromium.org/blink/trunk@198983 bbb929c8-8fbe-4397-9dbb-9b2b20218538