Web MIDI: Add a LayoutTest to check midimessage event handler
Now, Web MIDI has a bug that addEventListener() can not change
the internal state to open implicitly.
This is a test to check if the handler work.

If onmidimessage is set directly, it works. But if a handler is set
via addEventListener('midimessage'), it does not.
In this test, open() is called explicitly as a workaround.

BUG=510246

Review URL: https://codereview.chromium.org/1237393003

git-svn-id: svn://svn.chromium.org/blink/trunk@199032 bbb929c8-8fbe-4397-9dbb-9b2b20218538