IndexedDB: Make custom IDBCursor->IDBRequest wrapper linking more robust
An IDBCursor's associated IDBRequest should be re-used each time the
cursor advances. To keep the request's wrapper alive a hidden
reference is set on the cursor's wrapper, with custom bindings code.

Move that code to a custom toV8() overload rather than repeating it in
all the places that might call toV8(), to simplify the code and make
it less fragile.

R=haraken@chromium.org
TEST=storage/indexeddb/cursor-request-cycle.html
TEST=storage/indexeddb/key-cursor-request-cycle.html

Review URL: https://codereview.chromium.org/1240703004

git-svn-id: svn://svn.chromium.org/blink/trunk@199141 bbb929c8-8fbe-4397-9dbb-9b2b20218538