Fix the clearing of pointers in CompositedDeprecatedPaintLayerMapping
A previous patch addressing the clearing of pointers on layers removed
from a grouped mapping was nulling too many pointers, resulting in crashes
in some cases and missing content in others. This patch makes sure we
do not clear the pointer on somethign that appears multiple times in
the list.

R=vollick
BUG=508062,508264

Review URL: https://codereview.chromium.org/1245973002

git-svn-id: svn://svn.chromium.org/blink/trunk@199229 bbb929c8-8fbe-4397-9dbb-9b2b20218538