LayoutTest for hit test cache overlay scrollbar bug
Add ability to show and hide the overlay scrollbars
dynamically.
    
Add ability to ask for a validated hit request as
the validation will be removed eventually.
    
Add a LayoutTest for the fix in r199052. We make the scrollbar
appear and then generate a hit test once it disappears.    

BUG=398920

Review URL: https://codereview.chromium.org/1240263003

git-svn-id: svn://svn.chromium.org/blink/trunk@199342 bbb929c8-8fbe-4397-9dbb-9b2b20218538