When removing the srcdoc attribute, load src if it is present.
The content of frames can be set with either the src attribute or the
or the srcdoc attribute. (Unlike the src attribute, which contains a
URL, srcdoc contains attribute-escaped HTML. Blink loads about:srcdoc
to direct the loader to process the value of the srcdoc attribute.)

If both attributes are present srcdoc takes precedence. However the
existing handling of srcdoc did not bother to check whether the
attribute was being set or removed. As a result, removing a srcdoc
attribute would display a blank document when it should fall back to
loading the URL in the src attribute.

This implements that fallback behavior.

BUG=513689

Review URL: https://codereview.chromium.org/1261573003

git-svn-id: svn://svn.chromium.org/blink/trunk@199567 bbb929c8-8fbe-4397-9dbb-9b2b20218538