Fix invalidation of layers in grouped mappings
Previous patches cleaning up pointer re-use moved invalidation of
the layer for a change in squashing status from one layer to another.
This was a bug, leading to missed invalidation on commonly use sites
and in inspector.
    
As it turns out, the existing add-remove-squashed-layers.html test
was indeed testing correctly, and the change in that test earlier
should have been investigated further. Now the test is back to its
original result.

R=pdr@chromium.org,vollick@chromium.org
BUG=509538

Review URL: https://codereview.chromium.org/1255973005

git-svn-id: svn://svn.chromium.org/blink/trunk@199661 bbb929c8-8fbe-4397-9dbb-9b2b20218538