Fix root layer overflow clipping when viewport > initial containing block.
This addresses Android underpainting issues observed with --root-layer-scrolls
when the minimum page scale is < 1.  The changes are:

(1) We override LayoutBox::overflowClipRect in LayoutView to use the size of the
    layout viewport instead of the LayoutView's box size.

(2) We set the size of CompositedDeprecatedPaintLayerMapping::m_scrollingLayer
    based on the overflowClipRect instead of the clientBoxRect.  This is the
    GraphicsLayer that clips the scrolling content.

(3) DeprecatedPaintLayerClipper sets backgroundRect based on the viewport size.
    This ensures that the PaintInfo object has the correct dirty rect for
    background painting.

CSS overflow:overlay has the bizarre effect of making a scrollbar "overlay" the
scrolling content even when it is not actually an overlay scrollbar.  In this
edge case, m_scrollingLayer now excludes the space occupied by the scrollbars.
There is no observable difference, but a layout test that uses overflow:overlay
and dumps internals.layerTreeAsText has its expectations updated accordingly.

BUG=492871

Review URL: https://codereview.chromium.org/1262063002

git-svn-id: svn://svn.chromium.org/blink/trunk@199686 bbb929c8-8fbe-4397-9dbb-9b2b20218538