Removing dependency on SkDeferredCanvas from blink.
This change refactors Canvas2DLayerBridge to leverage the overdraw
optimization logic that is now built-in to SkCanvas (for copy on write)
and CanvasRenderingContext2D (for frame skipping), thus rendering
SkDeferredCAnvas unnecessary. A side-effect of this refactor it
replaces SkBitmaps with SkImages in many places.

The rebaselines are imperceptible pixel differences due to the slightly
differently formulated arithmetic in the StaticBitmapImage/SkImage
code paths. Use cases that resize images seem to be the most sensitive
to this change.

This change removes the Canvas2DLayerManager completely, so we will
need to follow-up with automatic flushing heuristics similar to
what is in RecordingImageBufferSurface in order to keep a cap on
resource consumption.

BUG=512991

Review URL: https://codereview.chromium.org/1259763002

git-svn-id: svn://svn.chromium.org/blink/trunk@199900 bbb929c8-8fbe-4397-9dbb-9b2b20218538