Reserve capacity API for hash table based containers
Before the change the WTF hash table based containers did not have API that would allow to reserve capacity for the known elements count
(insufficient hash table capacity forces the table to be reallocated and all the contained items to be rehashed several times while the container is being populated).
The only workaround was providing traits with the modified 'minimumTableSize' value but this turned to be inconvenient as
1) requires writing too much code on the client side just for setting the capacity
2) allows set the capacity only once
3) unnesesarily modifies the container type (i.e. HashSet<StringImpl*> != HashSet<StringImpl*, DefaultHash<StringImpl*>, UpdatedTraits>)

This patch adds new 'reserveCapacityForSize(unsigned size)' method to HashTable and HashSet classes in order to solve the issues described above.

The new API is used in blink::createHtmlCaseInsensitiveAttributesSet() function.

Review URL: https://codereview.chromium.org/1268893003

git-svn-id: svn://svn.chromium.org/blink/trunk@199971 bbb929c8-8fbe-4397-9dbb-9b2b20218538