Re-Relanding 'Always notify the WebMediaPlayer of any seek' patch
This change removes short circuit logic for seeking to the current position.
There are certain situations, like seeking to a truncated duration, where the
HTMLMediaElement is not in the best position to determine whether a call to
the WebMediaPlayer is necessary or not. This change allows the
WebMediaPlayer implementation to determine what the best course
of action should be. This also makes the seeking behavior more spec
compliant because even a seek to the current position should have the
seeking attribute be true until a stable state occurs. This is guaranteed
when calling into the WebMediaPlayer and allows all seeks to be treated the
same at the HTMLMediaElement level.

BUG=266631
TEST=LayoutTests/media/video-seek-to-current-position.html

Original CL: https://codereview.chromium.org/431903003/
Re-attemp: https://codereview.chromium.org/456343002/

Review URL: https://codereview.chromium.org/1267953003

git-svn-id: svn://svn.chromium.org/blink/trunk@199977 bbb929c8-8fbe-4397-9dbb-9b2b20218538