mac: Use a placeholder string for the family name of the system font.
This CL is inspired by mmaxfield@apple.com and
http://trac.webkit.org/changeset/184353.

The system font varies by OS version, as well as the version of the OSX SDK that
the binary is compiled against. This CL uses a placeholder string for the family
name of the system font that is passed around.

This CL has two positive effects.

1) Logic in LayoutTestControl* was checking the family name of the font directly
against the string "Lucida Grande". This was broken any time the system font
wasn't "Lucida Grande".
2) Layout tests were failing on OSX 10.10, 10.11, and when the binary was
compiled against the OSX 10.10 SDK. This CL fixes this particular set of
problems by always using the font "Lucida Grande" for layout tests.

This CL has one non-regression behavior change. If the font family of a text
control is set to "Lucida Grande", its width is no longer modified by the
special logic in LayoutTestControl*. This isn't a regression because there is no
reason to expect the special logic of LayoutTestControl to apply to the text
control to begin with (since "Lucida Grande" may or may not be the system font).
If we really want the system font to always have this behavior, we should
implement font family keywords like "-apple-system". See 515989 for more
details.

BUG=515989, 515836

Review URL: https://codereview.chromium.org/1266403002

git-svn-id: svn://svn.chromium.org/blink/trunk@199998 bbb929c8-8fbe-4397-9dbb-9b2b20218538