Let the WebMediaPlayer decide whether to use overlay video.
Previously, the runtime feature "overlayFullscreenVideoEnabled" was used
to decide whether to do fullscreen video overlay. This was set on
Android and for layout tests.
Now the WebMediaPlayer decides whether overlay fullscreen mode will be used.
This enables support for WebMediaPlayerImpl on Android, which will use
the standard composited fullscreen path.
The runtime feature is renamed to forceOverlayFullscreenVideo and retained
as an override for use in layout tests.

BUG=511376

Review URL: https://codereview.chromium.org/1254613003

git-svn-id: svn://svn.chromium.org/blink/trunk@200005 bbb929c8-8fbe-4397-9dbb-9b2b20218538