Show "Service Worker Fallback" instead of "canceled" for Service Worker fallback response.
When a CORS fetch request goes from a page to the service worker and it doesn't handle the request,
the browser process returns "CORS fallback response" to the page's renderer process.
And the page's renderer sends the request again with skip_service_worker flag.
It is because CORS check logic is implemented in the renderer process not in the browser process.

This fallback response is shown as cancelled in DevTools Network grid. But it is incorrect.
So this cl changes it to show as "(Service Worker Fallback)".

Screenshot: https://crbug.com/511054#c9

BUG=511054

Review URL: https://codereview.chromium.org/1271603003

git-svn-id: svn://svn.chromium.org/blink/trunk@200020 bbb929c8-8fbe-4397-9dbb-9b2b20218538