More cleanup after r200095.
Add back a bunch of lines that turned out to be needed.
This is filling in the holes for the bots that aren't
covered by the CQ.

TBR=pkasting@chromium.org

Initial cleanup patch after r200095.

Add back a bunch of lines that turned out to be needed.
This is filling in the holes for the bots that aren't
covered by the CQ.

TBR=pkasting@chromium.org

Review URL: https://codereview.chromium.org/1277073002 .

git-svn-id: svn://svn.chromium.org/blink/trunk@200122 bbb929c8-8fbe-4397-9dbb-9b2b20218538