Reworks the clipping logic
In the old code clipping calculation was done piece by piece storing
state information in a cache in each layer that was used by later
invocations of backgroundClipRects to avoid traveling all the way up
to the context.rootLayer each time. If the wrong value is stored in
the cache, it quietly traversed all the way to the root ignoring the
cache.

In the new code, a call to backgroundClipRect that is not a cache hit
results in a recursive precalculation that fills in all nodes using the
same root layer. This allows the cache to hold only one ClipRect per
ClipRectCacheSlot instead of three. It also clarifies the meaning of
root layer so that incorrect uses of rootLayer will not populate the
cache with bogus values.

The previous code could only store ClipRects relative to one node in
each cache, but this implementation allows two ClipRects to be passed
down the stack for the two possible root layers of any descendants.
In the old code, this resulted in cache misses and recalculated these
clips all the way to the root layer. The new code will therefore have
better performance in this case.

BUG=515970

Review URL: https://codereview.chromium.org/1242333003

git-svn-id: svn://svn.chromium.org/blink/trunk@200190 bbb929c8-8fbe-4397-9dbb-9b2b20218538