Make LayoutFlowThread::fragmentsBoundingBox() faster.
It used to let collectLayerFragments() handle the job, with an infinite dirty
rectangle. This forced us to collect all columns, when we really only needed to
know the bounding boxes of the first and the last columns in each fragmentainer
group.

This gives a major speed-up in the Layout/multicol/tall-content-short-columns-*
tests. It will also make it possible to submit a fix for bug 502407 and write a
test for it that doesn't time out.

R=jchaffraix@chromium.org,leviw@chromium.org

Review URL: https://codereview.chromium.org/1265933002

git-svn-id: svn://svn.chromium.org/blink/trunk@200264 bbb929c8-8fbe-4397-9dbb-9b2b20218538