Don't zoom in on a textbox if the user can't zoom back out.
If the textbox is contained in a node with a touch-action other than
auto or manipulation, the user can't pinch-zoom in or out over that
element, meaning the auto-zoom feature can trap a user in a zoomed in
state.

This fix looks at all the ancestors of the focused textbox. If any of them
disable pinch-zoom then we don't zoom in on the textbox since doing so may
leave the user zoomed in on a region that can't be easily zoomed out.

BUG=463197

Review URL: https://codereview.chromium.org/1274783003

git-svn-id: svn://svn.chromium.org/blink/trunk@200402 bbb929c8-8fbe-4397-9dbb-9b2b20218538