[tracing] Fix accounting of WebCache memory statistics
The 'other' field was not added to the memory stats returned by blink
WebCache. This CL fixes this issue, and also adds purged and purgeable
size to the stats. This will be used for memory-infra as a part of
chrome://tracing: see crrev.com/1283793004.

BUG=466141

Review URL: https://codereview.chromium.org/1278403007

git-svn-id: svn://svn.chromium.org/blink/trunk@200404 bbb929c8-8fbe-4397-9dbb-9b2b20218538