Move Position::isRenderedCharacter() for Text node to LayoutText
This patch moves an implementation of |Position::isRenderedCharacter()| for
|Text| node to |LayoutText| to reduce usage of |InlineBox| in |Position| class
for improving code health and helping identify layout API for editing.

This patch is a preparation of handling of :first-letter pseudo element in
selection.

Following patch will move other |InlineBox| references in editing/ to layout/.

BUG=174349, 518738
TEST=n/a; no behavior changes

Review URL: https://codereview.chromium.org/1285123002

git-svn-id: svn://svn.chromium.org/blink/trunk@200407 bbb929c8-8fbe-4397-9dbb-9b2b20218538