Move rendersInDifferentPosition() to VisibleUnits.cpp
This patch moves an |rendersInDifferentPosition()| and associated functions with
tests to "VisibleUnits.cpp" from "Position.cpp" since it is used only in
editing functions rather than dom to limit scope of the function for improving
code health.

This patch is a part of cleanup |Position| class.

BUG=518738
TEST=n/a; no behavior changes

Review URL: https://codereview.chromium.org/1290973002

git-svn-id: svn://svn.chromium.org/blink/trunk@200458 bbb929c8-8fbe-4397-9dbb-9b2b20218538