Makes V8LazyEventListener not throw an exception even if the script is wrong.
Not quite sure what behavior is expected with regard to the spec.
But Firefox and IE don't throw an exception and return null.
Thus, let's return undefined instead of throwing an exception.

In order to let developers realize an issue, the error message
is shown on the developer console.

BUG=514775

Review URL: https://codereview.chromium.org/1278533003

git-svn-id: svn://svn.chromium.org/blink/trunk@200477 bbb929c8-8fbe-4397-9dbb-9b2b20218538