WebCL: Move WebCLException to core/webcl.
This solves a layering violation, even if in an ugly way: ExceptionState
depends on V8ThrowException, which was including and depending on
WebCLException. The former is in bindings/core, which is not supposed to
depend on anything in modules/, which is where the latter was.

The WebCLException.* files do not have any dependency on other files in
modules/webcl (some unnecessary includes have been removed), so we can
easily move them to core/webcl instead and resolve the layering
violations.

The deeper issue of whether WebCLException should inherit from
DOMException and how to do this cleanly in Blink persists.

This change is a requirement for moving to M45, as work on upstream bug
358074 has reinforced these layering separations and causing code built
with component=shared_library to fail to link.