Merge updatePostLifecycleData into updateAllLifecyclePhasesInternal.
Having this in a separate method just encourages duplicate checks like
the NULL check here, and the local root frame check which are both not
needed. Move this into the only caller so the sequence of steps when
updating the lifecycle phases is more clear.

Review URL: https://codereview.chromium.org/1288613004

git-svn-id: svn://svn.chromium.org/blink/trunk@200584 bbb929c8-8fbe-4397-9dbb-9b2b20218538