Make CSSAnimationUpdate stack-allocated
This is a CL to fix (a part of) the regressions observed in tough_animation_cases
benchmarks on oilpan builds. In short, this CL fixes the heavy allocation design
of CSSAnimationUpdate.

Currently, a lot of CSSAnimationUpdate objects are allocated on Oilpan's heap.
This is problematic because the CSSAnimationUpdate objects hold HeapVectors and HeapHashMaps,
which produce a lot of unnecessary garbage in Oilpan's heap.

This CL makes CSSAnimationUpdate stack-allocated and explicitly clears the HeapVectors and HeapHashMaps
when the CSSAnimationUpdate is destructed on stack. When we need to hold a pending CSSAnimationUpdate
on CSSAnimation, it is held as a part of object of CSSAnimation. That way we can avoid
allocating unecessary HeapVectors, HeapHashMaps etc every time a pending CSSAnimationUpdate is created.

BUG=325467

Review URL: https://codereview.chromium.org/1281493004

git-svn-id: svn://svn.chromium.org/blink/trunk@200885 bbb929c8-8fbe-4397-9dbb-9b2b20218538