Abort forwarding remote input events when layoutObject is gone.
It is possible for an input event that will be forwarded to a remote
frame to also cause the iframe element in the local renderer to become
hidden, which destroys the layoutObject for that element. This was
causing a crash because the forwarding code requires a layoutObject for
mouse event conversion.

This patch causes event forwarding to abort in that situation.

No test because we don't yet have layout tests running with
--site-per-process.

BUG=520705
R=dcheng@chromium.org

Review URL: https://codereview.chromium.org/1298973002

git-svn-id: svn://svn.chromium.org/blink/trunk@200904 bbb929c8-8fbe-4397-9dbb-9b2b20218538