Make TextIterator to stop when it falls to handle shadow tree
The issue 521655 causes bad case in |toShadowRoot()| with a test script in
|TextIterator::advance()|. However, I could not get a HTML fragment causing
this situation since the test script generates HTML fragment and I could not
reproduce on my machines.

This patch changes |TextIterator| to stop when it fails to handle shadow tree
to avoid bad cast for preventing attacker to use this.

We'll add a test case for this once we have HTML fragment to cause this
bad cast.

BUG=521655
TEST=n/a; It is hard to create a test case for this issue

Review URL: https://codereview.chromium.org/1293703006

git-svn-id: svn://svn.chromium.org/blink/trunk@200958 bbb929c8-8fbe-4397-9dbb-9b2b20218538